requirements.txt
pytesseract
opencv-python
Pillow
easyocr
